Fix issues

+++ src/services/adminService.js
@@ -1008,6 +1008,9 @@ async function handleUnsubscribe(nonSharedToken, sharedToken, orgID, appID, apiR
     try {
         await sequelize.transaction(async (t) => {
 
+            console.log("Non Shared Token", sharedToken);
+            console.log("Non Shared Token", nonSharedToken);
+
             if (nonSharedToken.length > 0) {
                 await adminDao.deleteAppKeyMapping(orgID, appID, apiRefID);
             }
@@ -1022,7 +1025,9 @@ async function handleUnsubscribe(nonSharedToken, sharedToken, orgID, appID, apiR
                     tokenType: constants.TOKEN_TYPES.OAUTH
                 });
             } else {
-                await adminDao.deleteAppKeyMapping(orgID, appID, apiRefID, t);
+                if (sharedToken.length > 1) {
+                    await adminDao.deleteAppKeyMapping(orgID, appID, apiRefID, t);
+                }
             }
             await adminDao.deleteSubscription(orgID, subID, t);
         });
